requirements/test.txt
# Test dependencies go here.
-r base.txt
coverage==4.0b2
flake8==2.4.1
django-test-plus==1.0.9
factory_boy==2.5.2
mock-django==0.6.9
